python_backend_OBSOLETE/requirements.txt
certifi
click
flask
flask_mysql_connector
flask_cors
itsdangerous
jinja2
markupsafe
pika
pip
pyyaml
setuptools
vc
werkzeug
wheel
wincertstore
pyjwt